requirements.txt
tk
python3-discogs-client